Still fucked but let's simplify it
